refactoring: all UI classes are replaced in UI package
