+++ app/src/main/java/com/fictadvisor/android/ui/ResetPasswordFragment.kt
@@ -1,20 +1,18 @@
-package com.fictadvisor.android
+package com.fictadvisor.android.ui
 
-import android.net.Uri
 import android.os.Bundle
 import android.util.Log
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
+import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.Navigation
 import androidx.navigation.fragment.navArgs
 import com.fictadvisor.android.data.dto.AuthLoginResponse
 import com.fictadvisor.android.data.dto.BaseResponse
 import com.fictadvisor.android.data.dto.ResetPasswordDTO
-import com.fictadvisor.android.databinding.FragmentForgotPasswordBinding
 import com.fictadvisor.android.databinding.FragmentResetPasswordBinding
 import com.fictadvisor.android.repository.AuthRepository
 import com.fictadvisor.android.viewmodel.AuthViewModel
@@ -22,7 +20,6 @@ import com.fictadvisor.android.viewmodel.AuthViewModelFactory
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
-import okhttp3.ResponseBody
 
 class ResetPasswordFragment : Fragment() {
     private lateinit var binding: FragmentResetPasswordBinding
