+++ app/src/main/java/com/fictadvisor/android/ui/ScheduleFragment.kt
@@ -1,0 +1,31 @@
+package com.fictadvisor.android.ui
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.fictadvisor.android.R
+class ScheduleFragment : Fragment() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        arguments?.let {
+        }
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_schedule, container, false)
+    }
+
+    companion object {
+        @JvmStatic
+        fun newInstance(): ScheduleFragment {
+            return ScheduleFragment()
+        }
+
+    }
+}
