+++ app/src/main/java/com/fictadvisor/android/ui/WelcomeFragment.kt
@@ -1,11 +1,12 @@
-package com.fictadvisor.android
+package com.fictadvisor.android.ui
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import androidx.navigation.Navigation
+import com.fictadvisor.android.R
 import com.fictadvisor.android.databinding.FragmentWelcomeBinding
 
 class WelcomeFragment : Fragment() {
