+++ app/src/main/res/layout/fragment_continue_registration.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ContinueRegistrationFragment">
+    tools:context=".ui.ContinueRegistrationFragment">
 
     <!-- TODO: Update blank fragment layout -->
