+++ app/src/main/res/layout/fragment_registration.xml
@@ -5,7 +5,7 @@
     android:id="@+id/spinner_group"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RegistrationFragment">
+    tools:context=".ui.RegistrationFragment">
 
     <TextView
         android:id="@+id/textViewLogin1"
