+++ app/src/main/res/layout/fragment_schedule.xml
@@ -3,7 +3,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context=".ScheduleFragment">
+             tools:context=".ui.ScheduleFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
