+++ app/src/main/res/layout/fragment_verify_email.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".VerifyEmailFragment">
+    tools:context=".ui.VerifyEmailFragment">
 
     <TextView
         android:id="@+id/messageTV"
