+++ app/src/main/res/layout/fragment_welcome.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WelcomeFragment">
+    tools:context=".ui.WelcomeFragment">
 
     <!-- TODO: Update blank fragment layout -->
